sregistry.client.delete: fixes
* Ensure that queries need to be for single images
* If image could not be deleted on remote, call bot.exit()
* Support added for force flag in sregistry.main.s3.delete
* Delete looks broken in sregistry.main.google_[storage,build].delete

+++ sregistry/main/s3/delete.py
@@ -23,8 +23,24 @@ def remove(self, image, force=False):
     try:
         _object = self.bucket.Object(uri)
         _object.load() # this throws an exception if the object does not exist! -> if delete() fails no exception is thrown...
-        _object.delete()
+    
+        # taken from main.registry.delete.remove
+        continue_delete = True
+        if force is False:
+            response = input("Are you sure you want to delete %s?" % uri)
+            while len(response) < 1 or response[0].lower().strip() not in "ynyesno":
+                response = input("Please answer yes or no: ")
+            if response[0].lower().strip() in "no":
+                continue_delete = False
+
+        if continue_delete is True:
+            _object.delete()
+        else:
+            bot.info("Delete cancelled.")
     except Exception as e:
         bot.error('Could not delete object {}: {}'.format(uri, str(e)))
+        return None
+
+    return image
